Addressing PR review
